fix: 🐛 concat shares for not filled backup
Instead of listing shares from the list of directory, use the attrib_ file that describe shares in the directory. The method used is equals to the method used to list files, so the shares are filled even in the unfilled backup.

Method in attribute_file and hosts change (that should be a breaking change). But as it's a fix and the method of the view don't change, the commit will not be marked at breaking.